% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsws-package.R
\docType{data}
\name{nsws}
\alias{nsws}
\title{National Surface Water Survey data}
\format{An object of class \code{list} of length 8.}
\source{
\url{https://archive.epa.gov/emap/archive-emap/web/html/els.html}
}
\usage{
nsws
}
\description{
A dataset containing water quality and lake characteristics for
United States lakes. The data from the western lake survey is
wholly contained in the wlsds4 table.
}
\keyword{datasets}
