% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{nsws_ls}
\alias{nsws_ls}
\title{nsws_ls}
\usage{
nsws_ls(version_id, ...)
}
\arguments{
\item{version_id}{character version id}

\item{...}{extra arguments passed to list.files}
}
\description{
nsws_ls
}
\examples{
\dontrun{
nsws_ls("1")
}
}
